requirements.txt
python-dotenv==1.0.1
requests==2.32.3
Jinja2==3.1.4
openai>=1.40.0

# FastAPI Server (optional - only needed for API service)
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
